pyproject.toml
[tool.poetry]
name = "bigBox"
version = "0.1.0"
description = ""
authors = ["tapiaw38 <tapiaw38@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.1"
psycopg2-binary = "^2.9.1"
Pillow = "^8.4.0"
django-restframework = "^0.0.1"
Markdown = "^3.3.4"
django-filter = "^21.1"
django-environ = "^0.7.0"
django-extensions = "^3.1.3"
ipdb = "^0.13.9"
djangorestframework = "^3.12.4"
argon2-cffi = "^21.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8-isort = "^4.1.1"
flake8-black = "^0.2.3"
flake8-django = "^1.1.2"
flake8-simplify = "^0.14.2"
black = "^21.9b0"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
